updated header for all and for scratch meme generation added a few logic terms

+++ react-meme-app/src/meme-components/MemeCard.js
@@ -2,6 +2,7 @@ import React, { useState } from 'react'
 import { db } from '../firebase'
 import './MemeCard.css'
 import DeleteOutlineOutlinedIcon from '@material-ui/icons/DeleteOutlineOutlined';
+// import GetAppIcon from '@material-ui/icons/GetApp';
 import firebase from 'firebase'
 import { Button } from '@material-ui/core';
 import { useClipboard } from 'use-clipboard-copy';
@@ -11,34 +12,50 @@ function MemeCard({meme,type,id}) {
   const clipboard = useClipboard();
   const copyLink = () => {
 
-    clipboard.copy( meme.data.memes.url);
+    clipboard.copy(meme.data?.memes?.url || meme.data?.memes?.data?.url || meme.data?.memeURL);
     setCopied(true);
   };
     const user=firebase.auth().currentUser;
-    function deleteMeme() {
-  db.collection("user-memes")
-    .doc(user?.uid)
-    .collection("memes")
-    .doc(id)
-    .delete()
+
+    // Delete Meme
+
+  function deleteMeme(id, storagePath) {
+  const userId = firebase.auth().currentUser?.uid;
+  if (!userId || !id || !storagePath) return;
+
+  const storageRef = firebase.storage().ref();
+  const fileRef = storageRef.child(storagePath);
+
+  
+  fileRef.delete()
     .then(() => {
-      //console.log(`Deleted meme with id ${id}`);
+    
+      return db
+        .collection("user-memes")
+        .doc(userId)
+        .collection("memes")
+        .doc(id)
+        .delete();
+    })
+    .then(() => {
+      console.log("Meme deleted from both storage and Firestore.");
     })
     .catch((error) => {
       console.error("Error deleting meme:", error);
     });
 }
+
+
     return (
         <div >
                {type === "savelist" && (
           <div className="meme_card">
-            {/* Fetching imageUrl from data-->memes-->url (This is user-saved meme)|| data-->memes-->data--> (This is user-created meme) */}
-            <img src={meme.data.memes.url || meme.data.memes.data.url} alt="Meme"/>
+            <img src={meme.data?.memes?.url || meme.data?.memes?.data?.url || meme.data?.memeURL}  loading="lazy" alt="Meme"/>
             <div className='buttons_d'>
-             <Button variant="outlined" color="secondary" onClick={deleteMeme}>Delete Meme<DeleteOutlineOutlinedIcon/></Button>
-           
-              <Button title="You created this meme" onClick={copyLink} variant="outlined" color="primary">{copied?'Url Copied!':'Get URL🔗'}</Button>
-            
+             <Button variant="outlined" color="secondary" onClick={() => deleteMeme(meme.id, meme.data?.storagePath)}>Delete Meme<DeleteOutlineOutlinedIcon/></Button>
+
+             <Button title="You created this meme" onClick={copyLink} variant="outlined" color="primary">{copied?'Url Copied!':'Get URL🔗'}</Button>
+         
             </div>
           </div>
         )}
